% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haR-package.R
\docType{package}
\name{haR-package}
\alias{haR}
\alias{haR-package}
\title{haR: Heterogeneous AutoRegressive (HAR) model}
\description{
Tools for estimating and forecasting using the Heterogeneous AutoRegressive (HAR) model by Corsi (2009) \doi{10.1093/jjfinec/nbp001} and the Partial-Variances Heterogeneous AutoRegressive model by Bollerslev et al. (2022) \doi{10.1016/j.jeconom.2021.04.013}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/daniGiro/harmod}
  \item Report bugs at \url{https://github.com/daniGiro/harmod/issues}
}

}
\author{
\strong{Maintainer}: Daniele Girolimetto \email{daniele.girolimetto@phd.unipd.it} (\href{https://orcid.org/0000-0001-9387-1232}{ORCID}) [funder]

}
\keyword{internal}
