% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{har_palettes}
\alias{har_palettes}
\title{Complete list of palettes}
\format{An object of class \code{list} of length 7.}
\usage{
har_palettes
}
\description{
Use \code{\link{har_palette}} to construct palettes of desired length.
}
\keyword{datasets}

