% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/haRibo.R
\docType{package}
\name{haRibo}
\alias{haRibo}
\alias{haRibo-package}
\title{haRibo}
\description{
haRibo
}
\details{
List of palettes inspired from color combinations of Haribo Gummy Bears and similar candies
}

