% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/palettes.R
\name{haribo_palette}
\alias{haribo_palette}
\title{A palette generator inspired by the colors of Haribo gummy bears}
\usage{
haribo_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{goldbears}, \code{goldbearsUS},  \code{konfekt},
\code{picoballa}}

\item{n}{Number of colors desired. If \code{type} is set to "discrete", this
value will be limited to the number of colors available in the palette.
If omitted, uses all colours.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.}
}
\value{
A vector of colours.
}
\description{
These are a handful of color palettes from Haribo gummy bear compositions.
}
\keyword{colors}

