% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{haribo_palettes}
\alias{haribo_palettes}
\title{Complete list of palettes}
\format{\preformatted{List of 4
 $ goldbears  : chr [1:6] "#D9030B" "#FF2E00" "#FEAE01" "#FFDF00" ...
 $ goldbearsUS: chr [1:5] "#FB0000" "#FEC601" "#FEE202" "#FAF3D7" ...
 $ konfekt    : chr [1:6] "#79391F" "#C1722E" "#F8D54B" "#E4DEC9" ...
 $ picoballa  : chr [1:7] "#DE3135" "#70F2FF" "#FFBA76" "#4B9438" ...
}}
\usage{
haribo_palettes
}
\description{
Use \code{\link{haribo_palette}} to construct palettes of desired length.
}
\keyword{datasets}

