% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/har.R
\name{har}
\alias{har}
\title{Heterogeneous AutoRegressive (HAR) model}
\usage{
har(
  y,
  X,
  lags = c(1, 5, 22),
  fh = 1,
  log = FALSE,
  log_correction = TRUE,
  intercept = TRUE,
  min_correction = FALSE,
  min_constant = 0.5
)
}
\arguments{
\item{y}{A numeric vector (a realized measure of the integrated volatility).}

\item{X}{Optionally, a numerical vector or matrix of external regressors,
which must have the same number of rows as y.}

\item{lags}{A numeric vector specifying the lags to be used in the HAR model. Default is \code{c(1, 5, 22)}.}

\item{fh}{A numeric vector specifying the forecast horizons. Usually, 1 denotes daily volatility
(no aggregation), \code{5} weekly volatility, \code{22} monthly volatility. Default is \code{1}.}

\item{log}{A logical value indicating whether to apply a logarithmic transformation to the data.
Default is \code{FALSE}.}

\item{log_correction}{If \code{TRUE}, mean forecasts and fitted values are produced.}

\item{intercept}{A logical value indicating whether to include an intercept term in the model.
Default is \code{TRUE}.}

\item{min_correction}{A logical value indicating whether to apply a minimum correction
to negative forecast values. Default is \code{FALSE}.}

\item{min_constant}{A numeric value representing the minimum constant for corrections when
\code{min_correction} is applied. Default is \code{0.5}.}
}
\value{
An S3 object of class \link{haR} containing various components with information
related to the HAR model and its forecasts.
}
\description{
Estimation and forecasting using the HAR model by Corsi (2009) \doi{doi:10.1093/jjfinec/nbp001}
and the PV-HAR model (Partial-Variances Heterogeneous AutoRegressive model) by Bollerslev et
al. (2022) \doi{doi:10.1016/j.jeconom.2021.04.013}.
}
\examples{
x <- rnorm(1000)
obj <- har(x)

}
\references{
Corsi, F., 2009. A simple approximate long-memory model of realized volatility.
Journal of Financial Econometrics 7, 174–196. \doi{doi:10.1093/jjfinec/nbp001}.

Bollerslev, T., Medeiros, M., Patton, A., Quaedvlieg, R., 2022.
From zero to hero: Realized partial (co)variances.
Journal of Econometrics 231, 348–360. \doi{doi:10.1093/jjfinec/nbp001}.
}
