% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{haribo_palettes}
\alias{haribo_palettes}
\title{Complete list of palettes}
\format{\preformatted{List of 6
 $ goldbears       : chr [1:6] "#D9030B" "#FF2E00" "#FEAE01" "#FFDF00" ...
 $ goldbearsUS     : chr [1:5] "#FB0000" "#FEC601" "#FEE202" "#FAF3D7" ...
 $ konfekt         : chr [1:6] "#79391F" "#C1722E" "#F8D54B" "#E4DEC9" ...
 $ picoballa       : chr [1:7] "#DE3135" "#70F2FF" "#FFBA76" "#4B9438" ...
 $ baerchenpaerchen: chr [1:6] "#C9631B" "#459B3A" "#C5B347" "#7C365A" ...
 $ tropifrutti     : chr [1:6] "#637624" "#B5A12E" "#B5A78C" "#B38122" ...
}}
\usage{
haribo_palettes
}
\description{
Use \code{\link{haribo_palette}} to construct palettes of desired length.
}
\keyword{datasets}

